Fix missing transitive dependency for the VNDK.
When the vndk is enabled (BOARD_VNDK_VERSION=current),
vendor libraries must only link against vendor variants
in the build system. This also means that they receive
restricted sets of headers and that automatic global
headers are restricted. This sometimes causes certain system
headers to not be included when they are not included by
the VNDK version of headers.

Test: building with BOARD_VNDK_VERSION=current finds undefined symbols
Bug: 69526027
Change-Id: Ib245b83ed0f5a1a03db0ea8e0cd56fd1a650020d